Removed some Redundant files and did some code Refactoring
