+++ YtCmntAnalysis/Analysis/urls.py
@@ -2,8 +2,7 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path('', views.test, name='test'),
     path('about/', views.about, name='about'),
-    path('analysis/',views.Analview.as_view(), name='analysis'),
+    path('',views.Analview.as_view(), name='analysis'),
     
 ]
