+++ YtCmntAnalysis/Analysis/views.py
@@ -1,7 +1,5 @@
-from django.shortcuts import render, HttpResponse
-import os
+from django.shortcuts import render
 from django.core.cache import cache
-import json
 from django.apps import apps
 from django.views import View
 from .ML.sent_analysis import SimpleYtCommentAnalyzer
@@ -10,20 +8,6 @@ from urllib.parse import urlparse, parse_qs
 from .forms import YouTubeUrlForm
 from googleapiclient.errors import HttpError
 
-
-# Create your views here.
-def test(request):
-    """myapp_config = apps.get_app_config("Analysis")
-    model = myapp_config.tweet_nlp_model
-    print(type(model))
-    obj = SimpleYtCommentAnalyzer(model, "QwreMeXlFoY")
-    obj.get_comments_and_sentiment_by_video_id()
-    temp = obj.get_summary()
-    pretty_data = json.dumps(temp, indent=4)
-    response = HttpResponse(content_type="text/plain")
-    response.content = pretty_data
-    return HttpResponse(response)"""
-    return render(request, "sample_test.html")
 
 def about(request):
     return render(request, "about.html")
