+++ YtCmntAnalysis/YtCmntAnalysis/urls.py
@@ -17,14 +17,10 @@ from django.contrib import admin
 from django.urls import path, include
 from django.conf.urls.static import static
 from django.conf import settings
-from Analysis.views import test
-from Analysis.views import about
 
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path("sentanalysis/", include("Analysis.urls")),
-    path("", test, name="test"),
-    path("about/", about, name="about"),
+    path("", include("Analysis.urls")),
 ]
 urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
